requirements.txt
aiogram==3.20.0
apscheduler==3.10.0
python-dotenv==1.1.0